requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
python-jose
redis
celery
stripe
httpx
pydantic
pydantic-settings
rq